repo_src/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
python-multipart # Required for FastAPI file uploads
openai # For OpenRouter compatibility using the OpenAI SDK
psycopg2-binary # Keep if you plan to support PostgreSQL, otherwise remove for pure SQLite 
anthropic
pyyaml
tiktoken
notion-client # For Notion data source
discord.py # For Discord data source
APScheduler # For scheduled data fetching
mcp[cli] # Model Context Protocol server SDK